+++ src/rules/emails.rs
@@ -130,6 +130,7 @@ impl RuleChecker for EmailsRule {
                             file_byte_range: msgstr.byte_range.clone(),
                         },
                         edits,
+                        safe: false,
                     });
                     self.new_diag(checker, Severity::Warning, "different emails")
                         .map(|d| {