+++ src/rules/functions.rs
@@ -136,6 +136,7 @@ impl RuleChecker for FunctionsRule {
                             file_byte_range: msgstr.byte_range.clone(),
                         },
                         edits,
+                        safe: false,
                     });
                     self.new_diag(checker, Severity::Warning, "different functions")
                         .map(|d| {